+++ my-app/src/App.jsx
@@ -1,0 +1,21 @@
+
+import Header from "./components/Header/Header";
+
+import Main from "./components/Main";
+
+import Footer from "./components/Footer";
+
+
+
+
+const App = () => {
+  return (
+    <>
+      <Header/>
+      <Main/>
+      <Footer/>      
+    </>
+  );
+}
+
+export default App;
