--- my-app/src/components/Footer.js
@@ -1,31 +1,0 @@
-import React from "react";
-
-
-function Footer() {
-    return (
-
-        <div className="jumbotron" id="footerEdit">
-
-            <p>hola x2</p>
-
-            <div id="footerIcons">
-                <a href="https://www.facebook.com/wildlifemanagementmx"><i class="fab fa-facebook" ></i></a>
-                <a href="/"><i class="fab fa-twitter"></i></a>
-                <a href="/"><i class="fas fa-envelope"></i></a>
-                <a href="https://www.instagram.com/wildlifemanagementmx/?hl=es-la"><i class="fab fa-instagram"></i></a>
-            </div>
-            
-            <div id="footterCopyright">
-
-                <p>© Copyright - A quien corresponfa por "nombre de la empresa"</p>
-            </div>
-
-        </div>
-
-
-
-
-    )
-
-}
-export default Footer;
