+++ my-app/src/components/Footer.jsx
@@ -1,0 +1,18 @@
+
+const Footer = () => {
+    return (
+
+        <div className="jumbotron" id="footerEdit">
+
+            
+            <div id="footterCopyright">
+
+                <p>© Copyright - A quien corresponfa por "nombre de la empresa"</p>
+            </div>
+
+        </div>
+
+    )
+
+}
+export default Footer;
