--- my-app/src/components/Main.js
@@ -1,47 +1,0 @@
-import React from 'react'
-
-function Main () {
-    return (
-        <div className="container-main">    
-            <div id="noticias">
-                <h2>Noticias</h2>
-                <h3>Title description, Dec 7, 2017</h3>
-                <div className="fakeimg">Fake Image</div>
-                <p>Some text..</p>
-                <p>
-                    Sunt in culpa qui officia deserunt mollit anim id est laborum
-                    consectetur adipiscing elit, sed do eiusmod tempor incididunt ut
-                    labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-                    nostrud exercitation ullamco.
-                </p>
-            </div>
-            <br></br>
-            <div id="nosotros">
-                <h2>TITLE HEADING</h2>
-                <h5>Title description, Sep 2, 2017</h5>
-                <div className="fakeimg">Fake Image</div>
-                <p>Some text..</p>
-                <p>
-                Sunt in culpa qui officia deserunt mollit anim id est laborum
-                consectetur adipiscing elit, sed do eiusmod tempor incididunt ut
-                labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-                nostrud exercitation ullamco.
-                </p>
-            </div>
-            <div id="contacto">
-                <h2>TITLE HEADING</h2>
-                <h5>Title description, Sep 2, 2017</h5>
-                <div className="fakeimg">Fake Image</div>
-                <p>Some text..</p>
-                <p>
-                Sunt in culpa qui officia deserunt mollit anim id est laborum
-                consectetur adipiscing elit, sed do eiusmod tempor incididunt ut
-                labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-                nostrud exercitation ullamco.
-                </p>
-            </div>      
-      </div>
-    )
-}
-
-export default Main;
