--- my-app/src/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import  JournalApp  from './JournalApp';
-import './styles/style.css' ;
-
-
-
-ReactDOM.render(
-    <JournalApp/>,
-  document.getElementById('root')
-);
-
-
