+++ my-app/src/index.jsx
@@ -1,0 +1,9 @@
+import ReactDOM from 'react-dom';
+import App  from './App';
+import './styles/style.css' ;
+
+
+
+ReactDOM.render(<App/>, document.getElementById('root'));
+
+
